+++ src/main/java/com/twilio/rest/studio/v2/FlowReader.java
@@ -5,12 +5,11 @@
  *       /       /
  */
 
-package com.twilio.rest.serverless.v1.service.environment;
+package com.twilio.rest.studio.v2;
 
 import com.twilio.base.Page;
 import com.twilio.base.Reader;
 import com.twilio.base.ResourceSet;
-import com.twilio.converter.DateConverter;
 import com.twilio.exception.ApiConnectionException;
 import com.twilio.exception.ApiException;
 import com.twilio.exception.RestException;
@@ -19,79 +18,20 @@ import com.twilio.http.Request;
 import com.twilio.http.Response;
 import com.twilio.http.TwilioRestClient;
 import com.twilio.rest.Domains;
-import org.joda.time.DateTime;
 
 /**
- * PLEASE NOTE that this class contains preview products that are subject to
- * change. Use them with caution. If you currently do not have developer preview
- * access, please contact help@twilio.com.
+ * PLEASE NOTE that this class contains beta products that are subject to
+ * change. Use them with caution.
  */
-public class LogReader extends Reader<Log> {
-    private final String pathServiceSid;
-    private final String pathEnvironmentSid;
-    private String functionSid;
-    private DateTime startDate;
-    private DateTime endDate;
-
-    /**
-     * Construct a new LogReader.
-     *
-     * @param pathServiceSid The SID of the Service to read the Log resource from
-     * @param pathEnvironmentSid The SID of the environment with the Log resources
-     *                           to read
-     */
-    public LogReader(final String pathServiceSid,
-                     final String pathEnvironmentSid) {
-        this.pathServiceSid = pathServiceSid;
-        this.pathEnvironmentSid = pathEnvironmentSid;
-    }
-
-    /**
-     * The SID of the function whose invocation produced the Log resources to read..
-     *
-     * @param functionSid The SID of the function whose invocation produced the Log
-     *                    resources to read
-     * @return this
-     */
-    public LogReader setFunctionSid(final String functionSid) {
-        this.functionSid = functionSid;
-        return this;
-    }
-
-    /**
-     * The date/time (in GMT, ISO 8601) after which the Log resources must have been
-     * created. Defaults to 1 day prior to current date/time..
-     *
-     * @param startDate The date and time after which the Log resources must have
-     *                  been created.
-     * @return this
-     */
-    public LogReader setStartDate(final DateTime startDate) {
-        this.startDate = startDate;
-        return this;
-    }
-
-    /**
-     * The date/time (in GMT, ISO 8601) before which the Log resources must have
-     * been created. Defaults to current date/time..
-     *
-     * @param endDate The date and time before which the Log resource must have
-     *                been created.
-     * @return this
-     */
-    public LogReader setEndDate(final DateTime endDate) {
-        this.endDate = endDate;
-        return this;
-    }
-
+public class FlowReader extends Reader<Flow> {
     /**
      * Make the request to the Twilio API to perform the read.
      *
      * @param client TwilioRestClient with which to make the request
-     * @return Log ResourceSet
+     * @return Flow ResourceSet
      */
     @Override
-    public ResourceSet<Log> read(final TwilioRestClient client) {
+    public ResourceSet<Flow> read(final TwilioRestClient client) {
         return new ResourceSet<>(this, client, firstPage(client));
     }
 
@@ -99,15 +39,15 @@ public class LogReader extends Reader<Log> {
      * Make the request to the Twilio API to perform the read.
      *
      * @param client TwilioRestClient with which to make the request
-     * @return Log ResourceSet
+     * @return Flow ResourceSet
      */
     @Override
     @SuppressWarnings("checkstyle:linelength")
-    public Page<Log> firstPage(final TwilioRestClient client) {
+    public Page<Flow> firstPage(final TwilioRestClient client) {
         Request request = new Request(
             HttpMethod.GET,
-            Domains.SERVERLESS.toString(),
-            "/v1/Services/" + this.pathServiceSid + "/Environments/" + this.pathEnvironmentSid + "/Logs",
+            Domains.STUDIO.toString(),
+            "/v2/Flows",
             client.getRegion()
         );
 
@@ -120,11 +60,11 @@ public class LogReader extends Reader<Log> {
      *
      * @param targetUrl API-generated URL for the requested results page
      * @param client TwilioRestClient with which to make the request
-     * @return Log ResourceSet
+     * @return Flow ResourceSet
      */
     @Override
     @SuppressWarnings("checkstyle:linelength")
-    public Page<Log> getPage(final String targetUrl, final TwilioRestClient client) {
+    public Page<Flow> getPage(final String targetUrl, final TwilioRestClient client) {
         Request request = new Request(
             HttpMethod.GET,
             targetUrl
@@ -141,12 +81,12 @@ public class LogReader extends Reader<Log> {
      * @return Next Page
      */
     @Override
-    public Page<Log> nextPage(final Page<Log> page,
-                              final TwilioRestClient client) {
+    public Page<Flow> nextPage(final Page<Flow> page,
+                               final TwilioRestClient client) {
         Request request = new Request(
             HttpMethod.GET,
             page.getNextPageUrl(
-                Domains.SERVERLESS.toString(),
+                Domains.STUDIO.toString(),
                 client.getRegion()
             )
         );
@@ -161,12 +101,12 @@ public class LogReader extends Reader<Log> {
      * @return Previous Page
      */
     @Override
-    public Page<Log> previousPage(final Page<Log> page,
-                                  final TwilioRestClient client) {
+    public Page<Flow> previousPage(final Page<Flow> page,
+                                   final TwilioRestClient client) {
         Request request = new Request(
             HttpMethod.GET,
             page.getPreviousPageUrl(
-                Domains.SERVERLESS.toString(),
+                Domains.STUDIO.toString(),
                 client.getRegion()
             )
         );
@@ -174,17 +114,17 @@ public class LogReader extends Reader<Log> {
     }
 
     /**
-     * Generate a Page of Log Resources for a given request.
+     * Generate a Page of Flow Resources for a given request.
      *
      * @param client TwilioRestClient with which to make the request
      * @param request Request to generate a page for
      * @return Page for the Request
      */
-    private Page<Log> pageForRequest(final TwilioRestClient client, final Request request) {
+    private Page<Flow> pageForRequest(final TwilioRestClient client, final Request request) {
         Response response = client.request(request);
 
         if (response == null) {
-            throw new ApiConnectionException("Log read failed: Unable to connect to server");
+            throw new ApiConnectionException("Flow read failed: Unable to connect to server");
         } else if (!TwilioRestClient.SUCCESS.apply(response.getStatusCode())) {
             RestException restException = RestException.fromJson(response.getStream(), client.getObjectMapper());
             if (restException == null) {
@@ -201,9 +141,9 @@ public class LogReader extends Reader<Log> {
         }
 
         return Page.fromJson(
-            "logs",
+            "flows",
             response.getContent(),
-            Log.class,
+            Flow.class,
             client.getObjectMapper()
         );
     }
@@ -214,18 +154,6 @@ public class LogReader extends Reader<Log> {
      * @param request Request to add query string arguments to
      */
     private void addQueryParams(final Request request) {
-        if (functionSid != null) {
-            request.addQueryParam("FunctionSid", functionSid);
-        }
-
-        if (startDate != null) {
-            request.addQueryParam("StartDate", startDate.toString());
-        }
-
-        if (endDate != null) {
-            request.addQueryParam("EndDate", endDate.toString());
-        }
-
         if (getPageSize() != null) {
             request.addQueryParam("PageSize", Integer.toString(getPageSize()));
         }