+++ src/main/java/com/twilio/twiml/voice/Gather.java
@@ -168,6 +168,22 @@ public class Gather extends TwiML {
         }
     }
 
+    public enum SpeechModel {
+        DEFAULT("default"),
+        NUMBERS_AND_COMMANDS("numbers_and_commands"),
+        PHONE_CALL("phone_call");
+
+        private final String value;
+
+        private SpeechModel(final String value) {
+            this.value = value;
+        }
+
+        public String toString() {
+            return value;
+        }
+    }
+
     private final List<Gather.Input> input;
     private final URI action;
     private final HttpMethod method;
@@ -184,6 +200,7 @@ public class Gather extends TwiML {
     private final Boolean bargeIn;
     private final Boolean debug;
     private final Boolean actionOnEmptyResult;
+    private final Gather.SpeechModel speechModel;
 
     /**
      * For XML Serialization/Deserialization
@@ -213,6 +230,7 @@ public class Gather extends TwiML {
         this.bargeIn = b.bargeIn;
         this.debug = b.debug;
         this.actionOnEmptyResult = b.actionOnEmptyResult;
+        this.speechModel = b.speechModel;
     }
 
     /**
@@ -271,6 +289,9 @@ public class Gather extends TwiML {
         }
         if (this.isActionOnEmptyResult() != null) {
             attrs.put("actionOnEmptyResult", this.isActionOnEmptyResult().toString());
+        }
+        if (this.getSpeechModel() != null) {
+            attrs.put("speechModel", this.getSpeechModel().toString());
         }
 
         return attrs;
@@ -435,6 +456,15 @@ public class Gather extends TwiML {
     }
 
     /**
+     * Specify the model that is best suited for your use case
+     *
+     * @return Specify the model that is best suited for your use case
+     */
+    public Gather.SpeechModel getSpeechModel() {
+        return speechModel;
+    }
+
+    /**
      * Create a new {@code <Gather>} element
      */
     public static class Builder extends TwiML.Builder<Builder> {
@@ -454,6 +484,7 @@ public class Gather extends TwiML {
         private Boolean bargeIn;
         private Boolean debug;
         private Boolean actionOnEmptyResult;
+        private Gather.SpeechModel speechModel;
 
         /**
          * Input type Twilio should accept
@@ -605,6 +636,14 @@ public class Gather extends TwiML {
          */
         public Builder actionOnEmptyResult(Boolean actionOnEmptyResult) {
             this.actionOnEmptyResult = actionOnEmptyResult;
+            return this;
+        }
+
+        /**
+         * Specify the model that is best suited for your use case
+         */
+        public Builder speechModel(Gather.SpeechModel speechModel) {
+            this.speechModel = speechModel;
             return this;
         }